+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and `Removed`.
 - Multi-user support (#16): You can now debloat/restore for any user of the phone (not only the primary user 0). 
    - `Multi user mode` setting (default to on for Android 5+) allowing to remove packages for all users ([a work profile is another user](https://developer.android.com/work/managed-profiles)) instead of only the selected user.
    - User switcher (picklist).
-- New themes! (#11) : light, dark and dracula. Dracula theme is now the new default theme. Themes can be changed from the settings. 
+- New themes! (#11) : light, dark and lupin. Lupin theme is now the new default theme. Themes can be changed from the settings.
+- Description field scrollbar: you can now scroll long description (#40).
 
 ## Fixed
 - [Regression] Unsafe packages can be deleted without enabling `expert mode`.